Simplify and add get_tx_status method
