+++ crates/common/src/rpc/mod.rs
@@ -1,5 +1,4 @@
 //! Common RPC crate provides helper methods that are needed in rpc servers
-pub mod da;
 use std::time::Duration;
 
 use futures::future::BoxFuture;
